% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_data.R
\docType{data}
\name{box_data}
\alias{box_data}
\title{Box data for box_plot, include Value, Level1 and Level2 columns.}
\format{
Dataframe: include Value, Level1 and Level2 columns.
}
\usage{
data(box_data)
}
\description{
Box data for box_plot, include Value, Level1 and Level2 columns.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(box_data)

# 3. View example data
box_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/BoxStat/
}
\author{
benben-miao
}
\keyword{datasets}
