% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord_data.R
\docType{data}
\name{chord_data}
\alias{chord_data}
\title{Chord data as example data for the chord_plot function.}
\format{
Dataframe: include columns (Samples, Set1, Set2, ...), rows (Gene1, Gene2, ...).
}
\usage{
data(chord_data)
}
\description{
Chord data as example data for the chord_plot function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(chord_data)

# 3. View example data
chord_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/ChordPlot/
}
\author{
benben-miao
}
\keyword{datasets}
