% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_heatmap_data.R
\docType{data}
\name{circos_heatmap_data}
\alias{circos_heatmap_data}
\title{Gene expression as example data for circos_heatmap function.}
\format{
Dataframe: include columns (Sample1, Sample2, ...), rows (Gene1, Gene2, ...).
}
\usage{
data(circos_heatmap_data)
}
\description{
Gene expression as example data for circos_heatmap function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(circos_heatmap_data)

# 3. View example data
circos_heatmap_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/CircosHeatmap/
}
\author{
benben-miao
}
\keyword{datasets}
