% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_data.R
\docType{data}
\name{deg_data}
\alias{deg_data}
\title{Differentially expressed genes (DEGs) for volcano_plot function.}
\format{
Dataframe: include columns (Genes, Log2FoldChange, Pvalue, Padj), rows (Genes).
}
\usage{
data(deg_data)
}
\description{
Differentially expressed genes (DEGs) for volcano_plot function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(deg_data)

# 3. View example data
deg_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/VolcanoPlot/
}
\author{
benben-miao
}
\keyword{datasets}
