% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_data2.R
\docType{data}
\name{deg_data2}
\alias{deg_data2}
\title{Differentially expressed genes (DEGs) for ma_plot function.}
\format{
Dataframe: include columns (Genes, baseMean, Log2FoldChange, Padj), rows (Genes).
}
\usage{
data(deg_data2)
}
\description{
Differentially expressed genes (DEGs) for ma_plot function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(deg_data2)

# 3. View example data
deg_data2

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/MversusA/
}
\author{
benben-miao
}
\keyword{datasets}
