% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs_lists.R
\docType{data}
\name{degs_lists}
\alias{degs_lists}
\title{Paired comparisons differentially expressed genes (degs) among groups.}
\format{
Dataframe: Paired comparisons differentially expressed genes (degs) among groups (1st-col~: degs of paired comparisons).
}
\usage{
data(degs_lists)
}
\description{
Paired comparisons differentially expressed genes (degs) among groups.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(degs_lists)

# 3. View example data
degs_lists

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/VennPlot/
}
\author{
benben-miao
}
\keyword{datasets}
