% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs_stats.R
\docType{data}
\name{degs_stats}
\alias{degs_stats}
\title{All DEGs of paired comparison CT-vs-LT12 stats dataframe.}
\format{
Dataframe: All DEGs of paired comparison CT-vs-LT12 stats dataframe (1st-col: Genes, 2nd-col: log2FoldChange, 3rd-col: Pvalue, 4th-col: FDR).
}
\usage{
data(degs_stats)
}
\description{
All DEGs of paired comparison CT-vs-LT12 stats dataframe.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(degs_stats)

# 3. View example data
degs_stats

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/VolcanoPlot/
}
\author{
benben-miao
}
\keyword{datasets}
