% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs_stats2.R
\docType{data}
\name{degs_stats2}
\alias{degs_stats2}
\title{All DEGs of paired comparison CT-vs-LT12 stats2 dataframe.}
\format{
Dataframe: All DEGs of paired comparison CT-vs-LT12 stats2 dataframe (1st-col: Gene, 2nd-col: baseMean, 3rd-col: Log2FoldChange, 4th-col: FDR).
}
\usage{
data(degs_stats2)
}
\description{
All DEGs of paired comparison CT-vs-LT12 stats2 dataframe.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(degs_stats2)

# 3. View example data
degs_stats2

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/MversusA/
}
\author{
benben-miao
}
\keyword{datasets}
