% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cluster_data.R
\docType{data}
\name{gene_cluster_data}
\alias{gene_cluster_data}
\title{Gene cluster data for gene_cluster_trend function.}
\format{
Dataframe: Gene cluster data for gene_cluster_trend function.
}
\usage{
data(gene_cluster_data)
}
\description{
Gene cluster data for gene_cluster_trend function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(gene_cluster_data)

# 3. View example data
gene_cluster_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/BoxStat/
}
\author{
benben-miao
}
\keyword{datasets}
