% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exp.R
\docType{data}
\name{gene_exp}
\alias{gene_exp}
\title{Gene Expression dataframe of RNA-Seq.}
\format{
Dataframe: gene expression dataframe with cols (samples) and rows (genes).
}
\usage{
data(gene_exp)
}
\description{
Gene Expression dataframe of RNA-Seq.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset gene_exp
data(gene_exp)

# 3. View gene_exp
gene_exp

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/CorPlot/
}
\author{
benben-miao
}
\keyword{datasets}
