% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exp2.R
\docType{data}
\name{gene_exp2}
\alias{gene_exp2}
\title{Gene Expression dataframe of RNA-Seq.}
\format{
Dataframe: gene expression dataframe with cols (samples) and rows (genes).
}
\usage{
data(gene_exp2)
}
\description{
Gene Expression dataframe of RNA-Seq.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset gene_exp2
data(gene_exp2)

# 3. View gene_exp2
gene_exp2

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/CorPlot/
}
\author{
wei dong
}
\keyword{datasets}
