% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R
\docType{data}
\name{gene_expression}
\alias{gene_expression}
\title{Gene Expression dataframe of RNA-Seq.}
\format{
Dataframe: gene expression dataframe (1st-col: Transcripts or Genes, 2nd-col~: Samples).
}
\usage{
data(gene_expression)
}
\description{
Gene Expression dataframe of RNA-Seq.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset gene_expression
data(gene_expression)

# 3. View gene_expression
gene_expression

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/CorPlot/
}
\author{
benben-miao
}
\keyword{datasets}
