% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression2.R
\docType{data}
\name{gene_expression2}
\alias{gene_expression2}
\title{Shared DEGs of all paired comparisons in all samples expression dataframe of RNA-Seq.}
\format{
Dataframe: Shared DEGs of all paired comparisons in all samples expression dataframe of RNA-Seq. (1st-col: Genes, 2nd-col~: Samples).
}
\usage{
data(gene_expression2)
}
\description{
Shared DEGs of all paired comparisons in all samples expression dataframe of RNA-Seq.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset
data(gene_expression2)

# 3. View gene_expression2
gene_expression2

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/CorPlot/
}
\author{
benben-miao
}
\keyword{datasets}
