% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression3.R
\docType{data}
\name{gene_expression3}
\alias{gene_expression3}
\title{Shared DEGs of all paired comparisons in all groups expression dataframe of RNA-Seq.}
\format{
Dataframe: Shared DEGs of all paired comparisons in all groups expression dataframe of RNA-Seq. (1st-col: Genes, 2nd-col~n-1-col: Groups, n-col: Pathways).
}
\usage{
data(gene_expression3)
}
\description{
Shared DEGs of all paired comparisons in all groups expression dataframe of RNA-Seq.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset
data(gene_expression3)

# 3. View gene_expression3
gene_expression3

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/CorPlot/
}
\author{
benben-miao
}
\keyword{datasets}
