% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_go_kegg.R
\docType{data}
\name{gene_go_kegg}
\alias{gene_go_kegg}
\title{GO and KEGG annotation of background genes.}
\format{
Dataframe: GO and KEGG annotation of background genes (1st-col: Genes, 2nd-col: biological_process, 3rd-col: cellular_component, 4th-col: molecular_function, 5th-col: kegg_pathway).
}
\usage{
data(gene_go_kegg)
}
\description{
GO and KEGG annotation of background genes.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(gene_go_kegg)

# 3. View example data
gene_go_kegg

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/GOenrichStat/
}
\author{
benben-miao
}
\keyword{datasets}
