% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_anno.R
\docType{data}
\name{go_anno}
\alias{go_anno}
\title{GO annotation (BP, CC, MF) of genes as template data.}
\format{
Dataframe: include columns (id, biological_process, cellular_component, molecular_function),  symbol ";" split GO terms.
}
\usage{
data(go_anno)
}
\description{
GO annotation (BP, CC, MF) of genes as template data.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(go_anno)

# 3. View example data
go_anno

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/GOenrichStat/
}
\author{
benben-miao
}
\keyword{datasets}
