% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_deg_fc.R
\docType{data}
\name{go_deg_fc}
\alias{go_deg_fc}
\title{GO deg (differentially expressed genes) and fc (fold change) template data.}
\format{
Dataframe: include columns (id, log2FC).
}
\usage{
data(go_deg_fc)
}
\description{
GO deg (differentially expressed genes) and fc (fold change) template data.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(go_deg_fc)

# 3. View example data
go_deg_fc

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/GOenrichStat/
}
\author{
benben-miao
}
\keyword{datasets}
