% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_group_data.R
\docType{data}
\name{heatmap_group_data}
\alias{heatmap_group_data}
\title{Heatmap group data for heatmap_group function.}
\format{
Dataframe: include level1 columns (Groups, Group1, Group2, ...), level2 columns (GeneID, Sample1, Sample2, ...).
}
\usage{
data(heatmap_group_data)
}
\description{
Heatmap group data for heatmap_group function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(heatmap_group_data)

# 3. View example data
heatmap_group_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/Heatmap/
}
\author{
benben-miao
}
\keyword{datasets}
