% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_anno.R
\docType{data}
\name{kegg_anno}
\alias{kegg_anno}
\title{KEGG annotation (kegg_pathway) of genes as template data.}
\format{
Dataframe: include columns (id, kegg_pathway),  symbol ";" split ko pathways.
}
\usage{
data(kegg_anno)
}
\description{
KEGG annotation (kegg_pathway) of genes as template data.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(kegg_anno)

# 3. View example data
kegg_anno

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/KEGGenrichBar/
}
\author{
benben-miao
}
\keyword{datasets}
