% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_deg_fc.R
\docType{data}
\name{kegg_deg_fc}
\alias{kegg_deg_fc}
\title{KEGG deg (differentially expressed genes) and fc (fold change) template data.}
\format{
Dataframe: include columns (id, log2FC).
}
\usage{
data(kegg_deg_fc)
}
\description{
KEGG deg (differentially expressed genes) and fc (fold change) template data.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(kegg_deg_fc)

# 3. View example data
kegg_deg_fc

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/KEGGenrichStat/
}
\author{
benben-miao
}
\keyword{datasets}
