% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_data.R
\docType{data}
\name{network_data}
\alias{network_data}
\title{Network data as example for network_plot function.}
\format{
Dataframe: include columns (source_node, target_node).
}
\usage{
data(network_data)
}
\description{
Network data as example for network_plot function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(network_data)

# 3. View example data
network_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/NetworkPlot/
}
\author{
benben-miao
}
\keyword{datasets}
