% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_degs.R
\docType{data}
\name{paired_degs}
\alias{paired_degs}
\title{Paired comparisons differentially expressed genes (degs) among groups.}
\format{
Dataframe: Paired comparisons differentially expressed genes (degs) among groups (1st-col~: degs of paired comparisons).
}
\usage{
data(paired_degs)
}
\description{
Paired comparisons differentially expressed genes (degs) among groups.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(paired_degs)

# 3. View example data
paired_degs

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/VennPlot/
}
\author{
benben-miao
}
\keyword{datasets}
