% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_analysis.R
\name{pca_analysis}
\alias{pca_analysis}
\title{PCA dimensional reduction analysis for RNA-Seq.}
\usage{
pca_analysis(sample_gene, group_sample)
}
\arguments{
\item{sample_gene}{Dataframe: All genes in all samples expression dataframe of RNA-Seq (1st-col: Genes, 2nd-col~: Samples).}

\item{group_sample}{Dataframe: Samples and groups for gene expression (1st-col: Samples, 2nd-col: Groups).}
}
\value{
Table: PCA dimensional reduction analysis for RNA-Seq.
}
\description{
PCA dimensional reduction analysis for RNA-Seq.
}
\examples{
# 1. Library package TOmicsVis
library(TOmicsVis)

# 2. Load example datasets
data(gene_expression)
head(gene_expression)

data(samples_groups)
head(samples_groups)

# 3. Default parameters
res <- pca_analysis(gene_expression, samples_groups)
head(res)

}
\author{
benben-miao
}
