% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_group_sample.R
\docType{data}
\name{pca_group_sample}
\alias{pca_group_sample}
\title{PCA data2 include groups and samples.}
\format{
Dataframe: PCA data2 include groups and samples.
}
\usage{
data(pca_group_sample)
}
\description{
PCA data2 include groups and samples.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset pca_group_sample
data(pca_group_sample)

# 3. View pca_group_sample
pca_group_sample

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/PCAplot/
}
\author{
benben-miao
}
\keyword{datasets}
