% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA dimensional reduction visualization for RNA-Seq.}
\usage{
pca_plot(
  sample_gene,
  group_sample,
  xPC = 1,
  yPC = 2,
  multi_shape = TRUE,
  point_size = 5,
  point_alpha = 0.8,
  text_size = 5,
  fill_alpha = 0.05,
  border_alpha = 0,
  sci_fill_color = "Sci_AAAS",
  legend_pos = "right",
  legend_dir = "vertical",
  ggTheme = "theme_publication"
)
}
\arguments{
\item{sample_gene}{Dataframe: All genes in all samples expression dataframe of RNA-Seq (1st-col: Genes, 2nd-col~: Samples).}

\item{group_sample}{Dataframe: Samples and groups for gene expression (1st-col: Samples, 2nd-col: Groups).}

\item{xPC}{Numeric: PC index at x axis. Default: 1, options: 1, 2, 3, ...}

\item{yPC}{Numeric: PC index at y axis. Default: 2, options: 2, 3, 4, ...}

\item{multi_shape}{Logical: groups as shapes. Default: TRUE, options: TRUE, FALSE.}

\item{point_size}{Numeric: PCA plot point size. Default: 5, min: 0.}

\item{point_alpha}{Numeric: point color alpha. Default: 0.80, min: 0.00, max: 1.00.}

\item{text_size}{Numeric: PCA plot annotation size. Default: 5, min: 0.}

\item{fill_alpha}{Numeric: ellipse fill color alpha. Default: 0.10, min: 0.00, max: 1.00.}

\item{border_alpha}{Numeric: ellipse border color alpha. Default: 0.10, min: 0.00, max: 1.00.}

\item{sci_fill_color}{Character: ggsci color pallet. Default: "Sci_AAAS", options: "Sci_AAAS", "Sci_NPG", "Sci_Simpsons", "Sci_JAMA", "Sci_GSEA", "Sci_Lancet", "Sci_Futurama", "Sci_JCO", "Sci_NEJM", "Sci_IGV", "Sci_UCSC", "Sci_D3", "Sci_Material".}

\item{legend_pos}{Character: legend position. Default: "right", options: "none", "left", "right", "bottom", "top".}

\item{legend_dir}{Character: legend director. Default: "vertical", options: "horizontal", "vertical".}

\item{ggTheme}{Character: ggplot2 theme. Default: "theme_publication", options: "theme_default", "theme_bw", "theme_gray", "theme_light", "theme_linedraw", "theme_dark", "theme_minimal", "theme_classic", "theme_void".}
}
\value{
Plot: PCA dimensional reduction visualization for RNA-Seq.
}
\description{
PCA dimensional reduction visualization for RNA-Seq.
}
\examples{
# 1. Library package TOmicsVis
library(TOmicsVis)

# 2. Load example datasets
data(gene_expression)
head(gene_expression)

data(samples_groups)
head(samples_groups)

# 3. Default parameters
pca_plot(gene_expression, samples_groups)

# 4. Set multi_shape = FALSE
pca_plot(gene_expression, samples_groups, multi_shape = FALSE)

# 5. Set sci_fill_color = "Sci_NPG", fill_alpha = 0.10
pca_plot(gene_expression, samples_groups, sci_fill_color = "Sci_NPG", fill_alpha = 0.10)

}
\author{
benben-miao
}
