% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_sample_gene.R
\docType{data}
\name{pca_sample_gene}
\alias{pca_sample_gene}
\title{PCA data1 include samples and genes.}
\format{
Dataframe: PCA data1 include samples and genes.
}
\usage{
data(pca_sample_gene)
}
\description{
PCA data1 include samples and genes.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset pca_sample_gene
data(pca_sample_gene)

# 3. View pca_sample_gene
pca_sample_gene

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/PCAplot/
}
\author{
benben-miao
}
\keyword{datasets}
