% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_groups.R
\docType{data}
\name{samples_groups}
\alias{samples_groups}
\title{Samples and groups for gene expression.}
\format{
Dataframe: Samples and groups for gene expression (1st-col: Samples, 2nd-col: Groups).
}
\usage{
data(samples_groups)
}
\description{
Samples and groups for gene expression.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset samples_groups
data(samples_groups)

# 3. View samples_groups
samples_groups

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/PCAplot/
}
\author{
benben-miao
}
\keyword{datasets}
