% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_data.R
\docType{data}
\name{survival_data}
\alias{survival_data}
\title{Survival data as example data for survival_plot function.}
\format{
Dataframe: survival record data (1st-col: Time, 2nd-col: Status, 3rd-col: Group).
}
\usage{
data(survival_data)
}
\description{
Survival data as example data for survival_plot function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(survival_data)

# 3. View example data
survival_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/SurvivalAnalysis/
}
\author{
benben-miao
}
\keyword{datasets}
