% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_cross_data1.R
\docType{data}
\name{table_cross_data1}
\alias{table_cross_data1}
\title{Table cross data1 for table_cross function.}
\format{
Dataframe: include multiple columns.
}
\usage{
data(table_cross_data1)
}
\description{
Table cross data1 for table_cross function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(table_cross_data1)

# 3. View example data
table_cross_data1

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/BoxStat/
}
\author{
benben-miao
}
\keyword{datasets}
