% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_cross_data2.R
\docType{data}
\name{table_cross_data2}
\alias{table_cross_data2}
\title{Table cross data2 for table_cross function.}
\format{
Dataframe: include multiple columns.
}
\usage{
data(table_cross_data2)
}
\description{
Table cross data2 for table_cross function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(table_cross_data2)

# 3. View example data
table_cross_data2

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/BoxStat/
}
\author{
benben-miao
}
\keyword{datasets}
