% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_filter.R
\name{table_filter}
\alias{table_filter}
\title{Table filter used to filter row by column condition.}
\usage{
table_filter(data, ...)
}
\arguments{
\item{data}{Dataframe: Length, Width, Weight, and Sex traits dataframe (1st-col: Value, 2nd-col: Traits, 3rd-col: Sex).}

\item{...}{Expression: multiple expressions.}
}
\value{
Table: table filter used to filter row by column condition.
}
\description{
Table filter used to filter row by column condition.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Use example dataset
data(traits_sex)
head(traits_sex)

# 3. Set height > 100 & eye_color == "black"
res <- table_filter(traits_sex, Sex == "Male" & Traits == "Weight" & Value > 40)
head(res)

}
\author{
benben-miao
}
