% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_filter_data.R
\docType{data}
\name{table_filter_data}
\alias{table_filter_data}
\title{Table filter data as example for table_filter function.}
\format{
Dataframe: include multiple columns and many rows.
}
\usage{
data(table_filter_data)
}
\description{
Table filter data as example for table_filter function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(table_filter_data)

# 3. View example data
table_filter_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/BoxStat/
}
\author{
benben-miao
}
\keyword{datasets}
