% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_merge_data.R
\docType{data}
\name{table_merge_data}
\alias{table_merge_data}
\title{Table merge data as example for table_merge function.}
\format{
Dataframe: include multiple variables.
}
\usage{
data(table_merge_data)
}
\description{
Table merge data as example for table_merge function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(table_merge_data)

# 3. View example data
table_merge_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/BoxStat/
}
\author{
benben-miao
}
\keyword{datasets}
