% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_split_data.R
\docType{data}
\name{table_split_data}
\alias{table_split_data}
\title{Table split data as example for table_split function.}
\format{
Dataframe: include columns (variables), rows (records).
}
\usage{
data(table_split_data)
}
\description{
Table split data as example for table_split function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(table_split_data)

# 3. View example data
table_split_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/TableCast/
}
\author{
benben-miao
}
\keyword{datasets}
