% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomicsvis.R
\name{tomicsvis}
\alias{tomicsvis}
\title{TOmicsVis shiny app start function.}
\usage{
tomicsvis()
}
\value{
Shinyapp: TOmicsVis shiny app.
}
\description{
TOmicsVis shiny app start function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

}
\author{
benben-miao
}
