% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits_sex.R
\docType{data}
\name{traits_sex}
\alias{traits_sex}
\title{Length, Width, Weight, and Sex traits dataframe.}
\format{
Dataframe: Length, Width, Weight, and Sex traits dataframe (1st-col: Value, 2nd-col: Traits, 3rd-col: Sex).
}
\usage{
data(traits_sex)
}
\description{
Length, Width, Weight, and Sex traits dataframe.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(traits_sex)

# 3. View example data
traits_sex

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/BoxStat/
}
\author{
benben-miao
}
\keyword{datasets}
