% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_data.R
\docType{data}
\name{trend_data}
\alias{trend_data}
\title{Trend data include genes expression in multiple traits as example for trend_plot function.}
\format{
Dataframe: include columns (Gene, Trait1, Trait2, ..., Pathway), rows (Genes).
}
\usage{
data(trend_data)
}
\description{
Trend data include genes expression in multiple traits as example for trend_plot function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(trend_data)

# 3. View example data
trend_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/ExpressionTrend/
}
\author{
benben-miao
}
\keyword{datasets}
