% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_data.R
\docType{data}
\name{tsne_data}
\alias{tsne_data}
\title{TSNE data as example for tsne_plot function.}
\format{
Dataframe: include columns (IDs, Trait1, Trait2, ...).
}
\usage{
data(tsne_data)
}
\description{
TSNE data as example for tsne_plot function.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(tsne_data)

# 3. View example data
tsne_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/TSNEplot/
}
\author{
benben-miao
}
\keyword{datasets}
