% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{validate_character_options}
\alias{validate_character_options}
\title{Validate character options}
\usage{
validate_character_options(x, param_name = "x", options)
}
\arguments{
\item{x}{Character: value to validate}

\item{param_name}{Character: parameter name for error message}

\item{options}{Character vector: allowed options}
}
\value{
Logical: TRUE if valid option (invisible)
}
\description{
Validate if character is one of allowed options
}
