% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{validate_hex_color}
\alias{validate_hex_color}
\title{Validate hex color code}
\usage{
validate_hex_color(x, param_name = "color_str")
}
\arguments{
\item{x}{Character: color string to validate}

\item{param_name}{Character: parameter name for error message}
}
\value{
Logical: TRUE if valid hex color
}
\description{
Validate if a string is a valid hex color code
}
\keyword{internal}
