% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{validate_is_dataframe_or_matrix}
\alias{validate_is_dataframe_or_matrix}
\title{Validate is data.frame or matrix}
\usage{
validate_is_dataframe_or_matrix(x, param_name = "x")
}
\arguments{
\item{x}{Object to validate}

\item{param_name}{Character: parameter name for error message}
}
\value{
Logical: TRUE if valid (invisible)
}
\description{
Validate input is a data.frame or matrix
}
