% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{validate_logical}
\alias{validate_logical}
\title{Validate logical value}
\usage{
validate_logical(x, param_name = "x")
}
\arguments{
\item{x}{Logical: value to validate}

\item{param_name}{Character: parameter name for error message}
}
\value{
Logical: TRUE if valid logical (invisible)
}
\description{
Validate if value is logical
}
