% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_data.R
\docType{data}
\name{venn_data}
\alias{venn_data}
\title{Venn data with multiple gene sets.}
\format{
Dataframe: include multiple gene sets.
}
\usage{
data(venn_data)
}
\description{
Venn data with multiple gene sets.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(venn_data)

# 3. View example data
venn_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/VennPlot/
}
\author{
benben-miao
}
\keyword{datasets}
