% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_sex.R
\docType{data}
\name{weight_sex}
\alias{weight_sex}
\title{Weight and Sex traits dataframe.}
\format{
Dataframe: Weight and Sex traits dataframe (1st-col: Weight, 2nd-col: Sex).
}
\usage{
data(weight_sex)
}
\description{
Weight and Sex traits dataframe.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(weight_sex)

# 3. View example data
weight_sex

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/QuantileQuantile/
}
\author{
benben-miao
}
\keyword{datasets}
