% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_gene_exp.R
\docType{data}
\name{wgcna_gene_exp}
\alias{wgcna_gene_exp}
\title{WGCNA gene expression data for wgcna_analysis.}
\format{
Dataframe: include gene expression data.
}
\usage{
data(wgcna_gene_exp)
}
\description{
WGCNA gene expression data for wgcna_analysis.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(wgcna_gene_exp)

# 3. View example data
wgcna_gene_exp

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/WGCNA/
}
\author{
benben-miao
}
\keyword{datasets}
