% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_sample_group.R
\docType{data}
\name{wgcna_sample_group}
\alias{wgcna_sample_group}
\title{WGCNA samples groups data for wgcna_analysis.}
\format{
Dataframe: include samples and groups data.
}
\usage{
data(wgcna_sample_group)
}
\description{
WGCNA samples groups data for wgcna_analysis.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(wgcna_sample_group)

# 3. View example data
wgcna_sample_group

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/WGCNA/
}
\author{
benben-miao
}
\keyword{datasets}
