% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{add_stat_annotation}
\alias{add_stat_annotation}
\title{Add statistical annotation to plot}
\usage{
add_stat_annotation(
  p,
  label,
  x = Inf,
  y = -Inf,
  hjust = 0,
  vjust = 0,
  size = 5
)
}
\arguments{
\item{p}{ggplot object: target plot}

\item{label}{Character: text label to add}

\item{x}{Numeric: x position. Default: Inf}

\item{y}{Numeric: y position. Default: -Inf}

\item{hjust}{Horizontal justification. Default: 0}

\item{vjust}{Vertical justification. Default: 0}

\item{size}{Text size. Default: 5}
}
\value{
Enhanced ggplot object
}
\description{
Add formatted statistical test results to plots
}
