% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord_plot.R
\name{chord_plot}
\alias{chord_plot}
\title{Chord plot for visualizing the relationships of pathways and genes.}
\usage{
chord_plot(
  data,
  multi_colors = "RainbowColors",
  color_alpha = 0.5,
  link_visible = TRUE,
  link_dir = -1,
  link_type = "diffHeight",
  sector_scale = "Origin",
  width_circle = 3,
  dist_name = 3,
  label_dir = "Vertical",
  dist_label = 0.3
)
}
\arguments{
\item{data}{Dataframe: include columns (Samples, Set1, Set2, ...), rows (Gene1, Gene2, ...).}

\item{multi_colors}{Character: color palette. Default: "RainbowColors", options: "VividColors", "RainbowColors".}

\item{color_alpha}{Numeric: color alpha. Default: 0.50, min: 0.00, max: 1.00.}

\item{link_visible}{Logical: links visible. Default: TRUE, options: TRUE, FALSE.}

\item{link_dir}{Numeric: links direction, use with link_type. Default: -1, options: -1, 0, 1, 2.}

\item{link_type}{Character: links type, use with link_dir. Default: "diffHeight", options: "diffHeight", "arrows".}

\item{sector_scale}{Character: sector scale method. Default: "Origin", options: "Origin", "Scale".}

\item{width_circle}{Numeric: outside circle width. Default: 3.0, min: 0.0, max: 10.0.}

\item{dist_name}{Numeric: the distance of name and circle. Default: 3.0, min: 0.0, max: 10.0.}

\item{label_dir}{Character: label director. Default: "Vertical", options: "Horizontal", "Vertical".}

\item{dist_label}{Numeric: the distance of label and circle. Default: 0.3, min: 0.0.}
}
\value{
Plot: chord plot is used to visualize complex relationships between samples and genes, as well as between pathways and genes.
}
\description{
Chord plot is used to visualize complex relationships between samples and genes, as well as between pathways and genes.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Use example dataset
data(chord_data)

# 3. Default parameters
chord_plot(chord_data)

# 4. Set multi_colors = "VividColors"
chord_plot(chord_data, multi_colors = "VividColors")

# 5. Set link_type = "arrows"
chord_plot(chord_data, link_type = "arrows")

# 6. Set sector_scale = "Scale"
chord_plot(chord_data, sector_scale = "Scale")

# 7. Set label_dir = "Horizontal"
chord_plot(chord_data, label_dir = "Horizontal")

}
\author{
benben-miao
}
