% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_heatmap.R
\name{circos_heatmap}
\alias{circos_heatmap}
\title{Circos heatmap plot for visualizing gene expressing in multiple samples.}
\usage{
circos_heatmap(
  data,
  low_color = "#0000ff",
  mid_color = "#ffffff",
  high_color = "#ff0000",
  gap_size = 10,
  cluster_method = "complete",
  distance_method = "euclidean",
  dend_height = 0.2,
  rowname_size = 0.8
)
}
\arguments{
\item{data}{Dataframe: include columns (Sample1, Sample2, ...), rows (Gene1, Gene2, ...).}

\item{low_color}{Character: min value color (color name or hex value). Default: "#0000ff".}

\item{mid_color}{Character: middle value color (color name or hex value). Default: "#ffffff".}

\item{high_color}{Character: high value color (color name or hex value). Default: "#ff0000".}

\item{gap_size}{Numeric: heatmap gap size. Default: 10, min: 0.}

\item{cluster_method}{Character: cluster methods. Default: "complete", options: "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid".}

\item{distance_method}{Character: distance methods. Default: "euclidean", options: "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski".}

\item{dend_height}{Numeric: dendgram height. Default: 0.20, min: 0.00, max: 0.50.}

\item{rowname_size}{Numeric: rowname font size. Default: 0.80, min: 0.10, max: 10.00.}
}
\value{
Plot: circos heatmap plot for visualizing gene expressing in multiple samples.
}
\description{
Circos heatmap plot for visualizing gene expressing in multiple samples.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Use example dataset
data(circos_heatmap_data)

# 3. Default parameters
circos_heatmap(circos_heatmap_data)

# 4. Set low_color = "#008800"
# circos_heatmap(circos_heatmap_data, low_color = "#008800")

# 5. Set gap_size = 5
# circos_heatmap(circos_heatmap_data, gap_size = 5)

# 6. Set dend_height = 0.40
# circos_heatmap(circos_heatmap_data, dend_height = 0.40)

}
\author{
benben-miao
}
