% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{get_ggsci_color}
\alias{get_ggsci_color}
\title{Get ggsci color scale by name}
\usage{
get_ggsci_color(pal_name = "Sci_AAAS", alpha = 1)
}
\arguments{
\item{pal_name}{Character: ggsci palette name}
}
\value{
Scale function or NULL
}
\description{
Get ggsci color scale function by palette name
}
\keyword{internal}
