% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{get_ggsci_fill}
\alias{get_ggsci_fill}
\title{Safe get ggsci fill scale}
\usage{
get_ggsci_fill(pal_name = "Sci_AAAS", alpha = 1)
}
\arguments{
\item{pal_name}{Character: palette name}

\item{alpha}{Numeric: transparency value. Default: 1}
}
\value{
Scale function or default
}
\description{
Safely get ggsci fill color scale with error handling
}
\keyword{internal}
