% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{get_ggtheme}
\alias{get_ggtheme}
\title{Get ggplot2 theme by name}
\usage{
get_ggtheme(theme_name = "theme_publication", base_size = 12)
}
\arguments{
\item{theme_name}{Character: theme name}

\item{base_size}{Numeric: base font size. Default: 12}
}
\value{
Theme object
}
\description{
Get ggplot2 theme object by theme name with enhanced customization
}
