% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_enrich_dot.R
\name{kegg_enrich_dot}
\alias{kegg_enrich_dot}
\title{KEGG enrichment analysis and dot plot (None/Exist Reference Genome).}
\usage{
kegg_enrich_dot(
  kegg_anno,
  kegg_deg_fc,
  padjust_method = "fdr",
  pvalue_cutoff = 1,
  qvalue_cutoff = 1,
  sign_by = "p.adjust",
  category_num = 30,
  font_size = 12,
  low_color = "#ff0000aa",
  high_color = "#008888aa",
  ggTheme = "theme_light"
)
}
\arguments{
\item{kegg_anno}{Dataframe: include columns (id, kegg_pathway),  symbol ";" split ko pathways.}

\item{kegg_deg_fc}{Dataframe: include columns (id, log2FC).}

\item{padjust_method}{Character: P-value adjust to Q-value. Default: "fdr" (false discovery rate), options: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{pvalue_cutoff}{Numeric: P-value cutoff. Recommend: small than 0.05.}

\item{qvalue_cutoff}{Numeric: Q-value cutoff. Recommend: small than 0.05.}

\item{sign_by}{Character: significant by. Default: "p.adjust", options: "pvalue", "p.adjust", "qvalue".}

\item{category_num}{Numeric: categories number to display. Default: 30, min: 1, max: NULL.}

\item{font_size}{Numeric: category font size. Default: 12.}

\item{low_color}{Character: low value (p-value or q-value) color (color name or hex value).}

\item{high_color}{Character: high value (p-value or q-value) color (color name or hex value).}

\item{ggTheme}{Character: ggplot2 themes. Default: "theme_light", options: "theme_default", "theme_bw", "theme_gray", "theme_light", "theme_linedraw", "theme_dark", "theme_minimal", "theme_classic", "theme_void"}
}
\value{
Plot: KEGG enrichment analysis and dot plot (None/Exist Reference Genome).
}
\description{
KEGG enrichment analysis and dot plot (None/Exist Reference Genome).
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Use example dataset
data(kegg_anno)
head(kegg_anno)

data(kegg_deg_fc)
head(kegg_deg_fc)

# 3. Default parameters
kegg_enrich_dot(kegg_anno, kegg_deg_fc)

# 4. Set padjust_method = "BH"
kegg_enrich_dot(kegg_anno, kegg_deg_fc, padjust_method = "BH")

# 5. Set category_num = 10
kegg_enrich_dot(kegg_anno, kegg_deg_fc, category_num = 10)

# 6. Set low_color = "#ff8800ff", high_color = "#008888ff"
kegg_enrich_dot(kegg_anno, kegg_deg_fc, low_color = "#ff8800ff", high_color = "#008888ff")

# 7. Set ggTheme = "theme_minimal"
kegg_enrich_dot(kegg_anno, kegg_deg_fc, ggTheme = "theme_minimal")

}
\author{
benben-miao
}
