% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_enrich_net.R
\name{kegg_enrich_net}
\alias{kegg_enrich_net}
\title{KEGG enrichment analysis and net plot (None/Exist Reference Genome).}
\usage{
kegg_enrich_net(
  kegg_anno,
  kegg_deg_fc,
  padjust_method = "fdr",
  pvalue_cutoff = 1,
  qvalue_cutoff = 1,
  category_num = 20,
  net_layout = "circle",
  net_circular = TRUE,
  low_color = "#ff0000aa",
  high_color = "#008800aa"
)
}
\arguments{
\item{kegg_anno}{Dataframe: include columns (id, kegg_pathway),  symbol ";" split ko pathways.}

\item{kegg_deg_fc}{Dataframe: include columns (id, log2FC).}

\item{padjust_method}{Character: P-value adjust to Q-value. Default: "fdr" (false discovery rate), options: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{pvalue_cutoff}{Numeric: P-value cutoff. Recommend: small than 0.05.}

\item{qvalue_cutoff}{Numeric: Q-value cutoff. Recommend: small than 0.05.}

\item{category_num}{Numeric: categories number to display. Default: 20, min: 1, max: NULL.}

\item{net_layout}{Character: network layout. Default: "circle", options: 'star', 'circle', 'gem', 'dh', 'graphopt', 'grid', 'mds', 'randomly', 'fr', 'kk', 'drl' or 'lgl'.}

\item{net_circular}{Logical: network circular. Default: TRUE, options: TRUE, FALSE.}

\item{low_color}{Character: low value (p-value or q-value) color (color name or hex value).}

\item{high_color}{Character: high value (p-value or q-value) color (color name or hex value).}
}
\value{
Plot: KEGG enrichment analysis and net plot (None/Exist Reference Genome).
}
\description{
KEGG enrichment analysis and net plot (None/Exist Reference Genome).
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Use example dataset
data(kegg_anno)
head(kegg_anno)

data(kegg_deg_fc)
head(kegg_deg_fc)

# 3. Default parameters
kegg_enrich_net(kegg_anno, kegg_deg_fc)

# 4. Set category_num = 10
kegg_enrich_net(kegg_anno, kegg_deg_fc, category_num = 10)

# 5. Set low_color = "#ff8800ff", high_color = "#008888ff"
kegg_enrich_net(kegg_anno, kegg_deg_fc, low_color = "#ff8800ff", high_color = "#008888ff")

}
\author{
benben-miao
}
