% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_data.R
\docType{data}
\name{network_data}
\alias{network_data}
\title{Network data from WGCNA tan module top-200 dataframe.}
\format{
Dataframe: Network data from WGCNA tan module top-200 dataframe (1st-col: Source, 2nd-col: Target).
}
\usage{
data(network_data)
}
\description{
Network data from WGCNA tan module top-200 dataframe.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(network_data)

# 3. View example data
network_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/NetworkPlot/
}
\author{
benben-miao
}
\keyword{datasets}
