% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_data.R
\docType{data}
\name{quantile_data}
\alias{quantile_data}
\title{Quantile data for quantile_plot example.}
\format{
Dataframe: Trait measurement of samples in multiple groups (1st-col: Values, 2nd-col: Groups).
}
\usage{
data(quantile_data)
}
\description{
Quantile data for quantile_plot example.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(quantile_data)

# 3. View example data
quantile_data

}
\references{
https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/QuantileQuantile/
}
\author{
benben-miao
}
\keyword{datasets}
