% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_filter.R
\name{table_filter}
\alias{table_filter}
\title{Table filter used to filter row by column condition.}
\usage{
table_filter(data, ...)
}
\arguments{
\item{data}{Dataframe: include multiple columns and many rows.}

\item{...}{Expression: multiple expressions.}
}
\value{
Table: table filter used to filter row by column condition.
}
\description{
Table filter used to filter row by column condition.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Use example dataset
data(table_filter_data)
head(table_filter_data)

# 3. Set height > 100 & eye_color == "black"
res <- table_filter(table_filter_data, height > 100 & eye_color == "black")
head(res)

# 4. Set eye_color == "blue" | eye_color == "yellow"
res <- table_filter(table_filter_data, eye_color == "blue" | eye_color == "yellow")
head(res)

}
\author{
benben-miao
}
