% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{validate_is_dataframe}
\alias{validate_is_dataframe}
\title{Validate is data.frame}
\usage{
validate_is_dataframe(x, param_name = "x")
}
\arguments{
\item{x}{Object to validate}

\item{param_name}{Character: parameter name for error message}
}
\value{
Logical: TRUE if valid (invisible)
}
\description{
Validate input is a data.frame, stop with error if not
}
\keyword{internal}
