% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_pipeline.R
\name{wgcna_pipeline}
\alias{wgcna_pipeline}
\title{WGCNA analysis pipeline for RNA-Seq.}
\usage{
wgcna_pipeline(wgcna_gene_exp, wgcna_sample_group)
}
\arguments{
\item{wgcna_gene_exp}{Dataframe: include gene expression data.}

\item{wgcna_sample_group}{Dataframe: include samples and groups data.}
}
\value{
WGCNA results in tempdir() directory of current session.
}
\description{
WGCNA analysis pipeline for RNA-Seq.
}
\examples{
# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Use example dataset
data(wgcna_gene_exp)
head(wgcna_gene_exp)

data(wgcna_sample_group)
head(wgcna_sample_group)

}
\author{
benben-miao
}
